Please enter the commit message for your changes. Lines starting     with '   ' will be ignored, and an empty message aborts the commit.
    On branch newest
    Changes to be committed:
   	modified:   Makefile
   	modified:   ba.h
   	modified:   bd.h
   	modified:   controller.cc
   	modified:   controller.h
   	modified:   decorator.h
   	modified:   dragon.cc
   	modified:   dragon.h
   	modified:   enemy.cc
   	modified:   enemy.h
   	modified:   entity.h
   	modified:   floor.cc
   	modified:   floor.h
   	new file:   layout1.txt
   	modified:   main.cc
   	modified:   merchant.h
   	new file:   myProgram
   	modified:   orc.h
   	modified:   perpotion.cc
   	modified:   player.cc
   	modified:   player.h
   	modified:   tempotion.h
   	new file:   test1.txt
   	modified:   textdisplay.cc
   	modified:   wa.h
   	modified:   wd.h

+++ bd.h
@@ -6,7 +6,7 @@
 class BD : public Tempotion{
 public:
     BD(Player* component) : Tempotion{ component }{}
-    int getDef() override{ return component->getDef() + 5; }
+    float getDef() override{ return component->getDef() + 5; }
     void use(Player*& p) override;
 };
