+++ wd.h
@@ -4,11 +4,12 @@
 #include "tempotion.h"
 #include "player.h"
 
-class WD : public Tempotion {
-    public:
-        WD(Player *component) : Tempotion{component} {}
-        int getDef() override { return component->getDef() - 5; }
-        void use(Player*& p) override;
+class WD : public Tempotion{
+public:
+    WD(Player* component) : Tempotion{ component }{}
+    float getDef() override{ return component->getDef() - 5; }
+
+    void use(Player*& p) override;
 };
 
 #endif
